improvements

+++ backend/src/controllers/movie.js
@@ -2,7 +2,18 @@ const { Movie } = require('../models');
 
 const updateMovieById = async (req, res) => {
   try {
+    const { userId } = res.locals;
     const { movieId } = req.params;
+
+    const updatingMovie = await Movie.findOne({
+      where: { id: movieId },
+      attribute: ['userId']
+    });
+
+    if (updatingMovie.userId !== userId) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+
     const [_, movie] = await Movie.update(
       req.body,
       {
@@ -12,14 +23,29 @@ const updateMovieById = async (req, res) => {
       }
     );
     res.json(movie);
-  } catch(error) {
+  } catch (error) {
     res.status(500).json(error);
   }
 }
 
 const deleteMovieById = async (req, res) => {
   try {
+    const { userId } = res.locals;
     const { movieId } = req.params;
+
+    const deletingMovie = await Movie.findOne({
+      where: { id: movieId },
+      attribute: ['userId']
+    });
+
+    if(!deletingMovie) {
+      return res.status(404).json({ message: 'Not found' })
+    }
+
+    if (deletingMovie.userId !== userId) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+
     await Movie.destroy({ where: { id: movieId } });
     const movies = await Movie.findAll({
       order: [
@@ -27,39 +53,60 @@ const deleteMovieById = async (req, res) => {
       ],
     });
     res.json(movies);
-  } catch(error) {
+  } catch (error) {
     res.status(500).json(error);
   }
 }
 
 const getMovieById = async (req, res) => {
   try {
+    const { userId } = res.locals;
     const { movieId } = req.params;
-    const [movie] = await Movie.findAll({ where: { id: movieId } });
+    const movie = await Movie.findOne({ where: { id: movieId } });
+    if(!movie) {
+      return res.status(404).json({ message: 'Not found' })
+    }
+    if (movie.userId !== userId) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
     res.json(movie);
-  } catch(error) {
+  } catch (error) {
     res.status(500).json(error);
   }
 }
 
 const getAllMovies = async (req, res) => {
   try {
+    const { userId } = res.locals;
     const movies = await Movie.findAll({
+      where: { userId },
       order: [
         ['release_date', 'DESC']
       ],
+      attributes: {
+        exclude: ['userId']
+      },
     });
     res.json(movies);
-  } catch(error) {
+  } catch (error) {
     res.status(500).json(error);
   }
 }
 
 const addMovie = async (req, res) => {
   try {
-    const movie = await Movie.create(req.body);
+    const { userId } = res.locals;
+    const movie = await Movie.create({
+      userId,
+      name: req.body.name,
+      rate: req.body.rate,
+      trailerSrc: req.body.trailerSrc,
+      description: req.body.description,
+      releaseDate: req.body.releaseDate,
+      thumbnailSrc: req.body.thumbnailSrc,
+    });
     res.json(movie);
-  } catch(error) {
+  } catch (error) {
     res.status(500).json(error);
   }
 }
